requirements.txt
gym==0.15.4
opencv-python
glfw
pyyaml
tensorboard
lxml
wandb